+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 2.6.4 (2017-10-19)
+
+* [fix] Uninstall ansi console before exiting.
+* [new] Automatically switch to basic prompt under CYGWIN.
+
 # Version 2.6.3 (2017-10-19)
 
 * [new] Add basic prompt mode (-DbasicPrompt) on generate goal, for the case where ConsoleUI doesn't work properly.